requirements.txt
python-dotenv
utils
openai
flask
gradio
langchain
tiktoken